WIP: otp initial hackery
First, make certain that "epmd" is running.  If not, "epmd &".

Start exactly one Corfu server.  The Erlang node naming scheme is
hard-coded right now, so running more than one won't work correctly.
This example assumes a server running in `-m` memory mode, so it's
automatically boostrapped.

Check that the server is visible from epmd:

    % epmd -names

... the output should include:

    name corfu at port {some TCP port number}

Then start an Erlang VM on the local machine.

    % erl -sname foo

Then send a message to the JVM Erlang node to query the layout server.
The name of the remote Erlang node will vary; in this example, it
is `'corfu@sfritchie-m01'`.

    > {cmdlet, 'corfu@sfritchie-m01'} ! {self(), make_ref(), ["corfu_layout", "query", "localhost:8000"]}.

Next, see the reply, which is sent to the shell process.

    > flush().

You should see something like:

    (foo@sfritchie-m01)194> flush().                                                Shell got {#Ref<0.0.5.1300>,
               ["OK",
                "{\n  \"layoutServers\": [\n    \"localhost:8000\"\n  ],\n  \"sequencers\": [\n    \"localhost:8000\"\n  ],\n  \"segments\": [\n    {\n      \"replicationMode\": \"CHAIN_REPLICATION\",\n      \"start\": 0,\n      \"end\": -1,\n      \"stripes\": [\n        {\n          \"logServers\": [\n            \"localhost:8000\"\n          ]\n        }\n      ]\n    }\n  ],\n  \"epoch\": 0\n}"]}
    ok