Fix real concurrency bug, yay!
See commit 4770d10c0f9b71e28e5808e5e168ac2170bec9f9 for EQC
counterexample details.

The problem is that NettyServerRouter's single check for an incoming
client's epoch # is outside of LayoutServer's 'synchronized' scheme.
So it's possible for a layout command such as prepare() to succeed
on the client side but send a reply with a new epoch number (due to
concurrent handling of a layout commit()).  When the client receives the
reply, the client side throws a bad epoch exception ... and then the
QuickCheck model believes that the operation failed when in fact the
server believes the op succeeded.

We also change NettyServerRouter's sendResponse() method to always send
a reply with the same epoch as the request.  If the new sanity check
fails, we log a warning level error and then exit the entire server to
express our displeasure.