Fix QC glue code for layout protocol's prepare() message.
NOTE: QC model is currently broken, will update in next commit.

A very nitpicky Dahlia asking lots of pesky questions made me
realize that the QC glue code for prepare() is omitting an
important part of the prepare()'s response: the server's
layout if that server had a prior propose()'d layout.

The glue code was only checking the boolean value of the
.isAccepted() method.