Fix two model problems in next_state
For propose: use the same success criteria for next state update
that postcondition() uses.  {sigh}  If I have a critique of this
style of QC rule writing, it's that it's too easy to keep what
ought to be duplicate conditionals in both functions in sync.

For prepare: do not reset the proposed layout here, see the
added comment.