Reproducing client for bug #273
Run server via:

    rm -rf /tmp/some/path && ./bin/corfu_server -l /tmp/some/path -s 9000 -d WARN

Run client via:

    bin/corfu_repro273 -c localhost:9000

When the bug is triggered, it appears as if one of the two threads
on the client side is permanently blocked trying to fill a hole.

However, sooner or later, the other thread will also be blocked.
I haven't investigated why.  The timing of when it stops appears a
bit random.  But the correlation between the two runtimes and trying
to write to different parts of the stream looks suspicious.