WIP: damn, weird, ExecutorService's invokeAll() isn't working as expected
Grrr.  I am expecting invokeAll() to return some damn futures,
regardless of how long it takes the damn computation to actually
run.

But if I use the coop scheduler (as I am in this WIP), and try
using invokeAll(), and try to *run* the scheduler after the
invokeAll(), ... I see:

* invokeall(no timeout) : never returns
* invokeAll(with timeout): only returns after timeout
    * as an unpleasant side-effect, the futures that return
      have 'cancelled' status, so their computation is lost.

Grrrrr, I'm going to have to rearrange things so that the coop
scheduler is running in a thread *prior* to this invokeAll() nonsense.