Target met: IDENTICAL = true, java command line invocation:
    java -javaagent:/Users/sfritchie/.m2/repository/org/aspectj/aspectjweaver/1.8.9/aspectjweaver-1.8.9.jar -classpath aspects/target/aspects-0.1-SNAPSHOT.jar:./test/target/test-classes:./infrastructure/target/infrastructure-0.1-SNAPSHOT-shaded.jar:$HOME/.m2/repository/org/aspectj/aspectjweaver/1.8.9/aspectjweaver-1.8.9.jar:$HOME/.m2/repository/junit/junit/4.12/junit-4.12.jar:$HOME/.m2/repository/commons-lang/commons-lang/2.6/commons-lang-2.6.jar:$HOME/.m2/repository/org/assertj/assertj-core/3.0.0/assertj-core-3.0.0.jar org.corfudb.runtime.checkpoint.CheckpointCoopTest

I have run out of time & patience to try to figure out how to
use the annotation-based AspectJ interface to be able to intercept
object instantiation & construction.  If I figure out later how
it can be done, then I'll use pure AspectJ and undo this commit.

As a workaround, I've made a small change to VersionLockedObject
to change the use of StampedLock -> CoopStampedLock.  The
CoopStampedLock will mimic being a StampedLock when used outside
of CoopScheduler.