Alt WIP: alternative method for triggering exactly GH 802
This is an alternative to triggering GH 802, using only a single
aspect pointcut to add a sched() after a getOptimisticStreamUnsafe()
call.  However, this isn't good enough: no NPE faults.

Only coop-thr-0 and coop-thr-1 are hitting the injected sched().
So, perhaps thread #2 is doing something to set our target pointer
to null, and we don't have a sched() in the proper place to permit
it.  So, we'll try adding a bit more.......